requirement.txt
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.10
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.0.1
django-reversion==3.0.0
docopt==0.6.2
djangorestframework==3.9.3
django-cors-headers=2.5.3
django-filter==2.1.0
djangorestframework-jwt==1.11.0
social-auth-app-django==3.1.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.11.3
idna==2.7
jdcal==1.4
mysqlclient==1.3.13
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.4
Pillow==5.2.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
wincertstore==0.2
xlrd==1.1.0
XlsxWriter==1.0.5
xlwt==1.3.0
yarg==0.1.9
